+++ database/migrations/2024_08_24_133455_create_payments_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('bill_id');
+            $table->unsignedBigInteger('bill_id');
             
             $table->timestamps();
